#56 Store queue of host chores on client side to avoid chore interrup… (#219)
#56 Store queue of host chores on client side to avoid chore interruption by host earlier completion.
