+++ src/MultiplayerMod/Game/Chores/ChoreConsumerEvents.cs
@@ -1,31 +1,37 @@
 using System;
+using HarmonyLib;
+using JetBrains.Annotations;
 using MultiplayerMod.Core.Logging;
 using MultiplayerMod.ModRuntime.Context;
+using MultiplayerMod.ModRuntime.StaticCompatibility;
 using MultiplayerMod.Multiplayer;
 using MultiplayerMod.Multiplayer.CoreOperations;
+using MultiplayerMod.Multiplayer.Objects;
+using MultiplayerMod.Multiplayer.Objects.Reference;
 
 namespace MultiplayerMod.Game.Chores;
 
-//[HarmonyPatch(typeof(ChoreConsumer), nameof(ChoreConsumer.FindNextChore))]
+[HarmonyPatch(typeof(ChoreConsumer), nameof(ChoreConsumer.FindNextChore))]
 public class ChoreConsumerEvents {
 
     private static readonly Core.Logging.Logger log = LoggerFactory.GetLogger<ChoreConsumerEvents>();
     public static event Action<FindNextChoreEventArgs>? FindNextChore;
 
-    // ReSharper disable once InconsistentNaming, UnusedMember.Global
+    [UsedImplicitly]
     [RequireMultiplayerMode(MultiplayerMode.Host)]
     [RequireExecutionLevel(ExecutionLevel.Game)]
-    public static void Postfix(ChoreConsumer __instance, Chore.Precondition.Context out_context, ref bool __result) {
+    public static void Postfix(ChoreConsumer __instance, Chore.Precondition.Context out_context, bool __result) {
         if (!__result)
             return;
 
-        var kPrefabID = __instance.gameObject.GetComponent<KPrefabID>();
-        var instanceId = kPrefabID.InstanceID;
+        if (Dependencies.Get<MultiplayerGame>().Mode != MultiplayerMode.Host)
+            return;
+
         var choreId = out_context.chore.id;
+        var reference = __instance.GetReference();
 
         var args = new FindNextChoreEventArgs(
-            instanceId,
-            __instance.ToString(),
+            reference,
             Grid.PosToCell(__instance.transform.position),
             choreId,
             out_context.chore.GetType(),
@@ -33,18 +39,15 @@ public class ChoreConsumerEvents {
             out_context.isAttemptingOverride
         );
         log.Debug(
-            $"Triggering {args.InstanceId} {args.InstanceString} {args.InstanceCell} {args.ChoreId} {args.ChoreType} {args.ChoreCell}"
+            $"Triggering {reference} {args.InstanceCell} {args.ChoreId} {args.ChoreType} {args.ChoreCell}"
         );
-        FindNextChore?.Invoke(
-            args
-        );
+        FindNextChore?.Invoke(args);
     }
 
 }
 
 public record FindNextChoreEventArgs(
-    int InstanceId,
-    string InstanceString,
+    ComponentReference<ChoreConsumer> Reference,
     int InstanceCell,
     int ChoreId,
     Type ChoreType,
