+++ src/MultiplayerMod/Multiplayer/Objects/Reference/MultiplayerIdReference.cs
@@ -17,4 +17,21 @@ public class MultiplayerIdReference : GameObjectReference {
 
     public override string ToString() => Id.ToString();
 
+    protected bool Equals(MultiplayerIdReference other)
+    {
+        return Id.Equals(other.Id);
+    }
+
+    public override bool Equals(object? obj)
+    {
+        if (ReferenceEquals(null, obj)) return false;
+        if (ReferenceEquals(this, obj)) return true;
+        return obj.GetType() == GetType() && Equals((MultiplayerIdReference)obj);
+    }
+
+    public override int GetHashCode()
+    {
+        return Id.GetHashCode();
+    }
+
 }
