+++ src/MultiplayerMod/Multiplayer/World/HostChores.cs
@@ -1,9 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using MultiplayerMod.Multiplayer.Objects.Reference;
 
 namespace MultiplayerMod.Multiplayer.World;
 
 public static class HostChores {
 
-    public static Dictionary<int, Queue<Chore.Precondition.Context>> Index { get; } = new();
+    public static Dictionary<ComponentReference<ChoreConsumer>, Queue<Func<Chore.Precondition.Context?>>> Index { get; } =
+        new();
 
 }
